userRegister, userLogin, userEdit, userDelte, Posts all tested with the backend

+++ src/components/UserDelete.jsx
@@ -7,17 +7,22 @@ export default function UserDelete(){
 
     function handleUserDelete() {
         console.log('user :>> ', user);
-        const config = {
-          credentials: 'include', // specify this if you need cookies
-          method: "DELETE",
-        };
-    
-        fetch(`http://localhost:7000/user/${user.id}`, config)
-          .then((response) => response.json())
-          .then((result) => console.log("UserRegistrationPOST:", result))
-          .catch((error) => console.log(error));
-        // Pop up message instead of console.log later
-        // cookie, Token stuff
+        if(user.id) {
+            const config = {
+                credentials: 'include', // specify this if you need cookies
+                method: "DELETE",
+              };
+              
+              fetch(`http://localhost:7000/user/${user.id}`, config)
+                .then((response) => response.json())
+                .then((result) => console.log("UserRegistrationPOST:", result))
+                .catch((error) => console.log(error));
+              // Pop up message instead of console.log later
+              // cookie, Token stuff
+        } else {
+            console.log("User is not logged in");
+        }
+
       }
 
     return(
